% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_averages.R
\name{allsky_monthly_qq}
\alias{allsky_monthly_qq}
\title{Calculate monthly all sky}
\usage{
allsky_monthly_qq(qq = qq_cc_day, frac = 0.8)
}
\arguments{
\item{qq}{daily global radiation observations}

\item{frac}{minimum fraction of daily observations in the month}
}
\value{
returns a dataframe with monthly values and additionally the season and year.
}
\description{
Needs only qq as input. A minimum of 80\% of the monthly values should be present
to calculate monthly averages.
}
\author{
Marieke Dirksen
}
