% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_averages.R
\name{clearsky_monthly_qq_cc}
\alias{clearsky_monthly_qq_cc}
\title{Calculate monthly clear sky}
\usage{
clearsky_monthly_qq_cc(qq_cc = qq_cc_day, min_obs = 2)
}
\arguments{
\item{min_obs}{minimum number of observations required to calculate monthly clear sky}

\item{qq_cc_day}{daily global radiation and cloud cover observations}
}
\value{
returns a dataframe with monthly values and additionally the season and year.
}
\description{
Needs both qq and cc as input. Clear sky monthly values are calculated if a 
minumum of two daily observations in the month with 0 or 1 okta were measured.
}
\author{
Marieke Dirksen
}
