% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_filling.R
\name{completeness.series}
\alias{completeness.series}
\title{Completeness of a time series}
\usage{
completeness.series(
  series,
  start = min(series$month_year),
  stop = max(series$month_year)
)
}
\arguments{
\item{series}{time series with STAID,month_year dates}

\item{start}{start of the time series or period}

\item{stop}{stop of the time series or period}
}
\description{
Calculates the fraction of measurements within the whole period of a series or other period
}
\author{
Marieke Dirksen
}
