% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_filling.R
\name{fill_gap}
\alias{fill_gap}
\title{Correlation with other stations}
\usage{
fill_gap(
  meta,
  series,
  full_data_series,
  missing_date,
  buffer_dist = 500,
  n.months = 5,
  n.stations = 5
)
}
\arguments{
\item{meta}{metadata with lat lon from all the stations}

\item{series}{series with missing data}

\item{full_data_series}{full data series which will be used to fill the gap}

\item{missing_date}{missing month from the series}

\item{buffer_dist}{minimum distance to the filling stations in km. Returning 1 if the full_data_series stations
falls within the buffer zone.}

\item{n.months}{minimum number of months with needed to compute correlation coef.}

\item{n.stations}{maximum number of reference stations from which the median is calculated}
}
\description{
calculates the correlation coefficient from the series with gaps to all the other
series and computes a buffer with a 500km radius. At least 5 available observations per month for
all the data is recommended, otherwise the normalization of the data and correlation coef. can
give a wrong impression.
}
\author{
Marieke Dirksen
}
