% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gap_filling.R
\name{fill_time_series}
\alias{fill_time_series}
\title{Fill gaps in time series}
\usage{
fill_time_series(staids, qq_meta, qq_monthly, fname)
}
\arguments{
\item{staids}{unique station ids corresponding with STAID in qq_monthly}

\item{qq_meta}{metadata from the stations are read by \link{read_ECAD_info}}

\item{qq_monthly}{monthly averaged file as calculated by either \link{allsky_monthly_qq} or \link{clearsky_monthly_qq_cc}}

\item{fname}{file name to which the filled gaps should be written}
}
\description{
Fill gaps in time series using the function \link{fill_gap}. The monthly
series which have the id column STAID are one by one selected, checked for gaps and
if possible filled. The filled values are written to a single file which is appended.
Additionally to the filling value the number of stations from which the median is calculated
and their standard deviation is included.
}
\author{
Marieke Dirksen
}
