% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_analysis.R
\name{get_change_point}
\alias{get_change_point}
\title{Find the change point in time-series}
\usage{
get_change_point(series, h.years = 10)
}
\arguments{
\item{series}{As calculated by \link{clearsky_monthly_qq_cc} or \link{allsky_monthly_qq}, possibly combined
with the gap filled series from \link{fill_time_series}.}

\item{h.years}{minimum number of years between the change point detections}
}
\value{
Returns the fitted trend model and a plot of the time series with trend and change points.
}
\description{
Calculates the change points and fits linear trends before and after the change points.
For the trend fitting a five year moving average window function was used to exclude short term variability
from the change point detection.
}
\author{
Marieke Dirksen
}
