% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_hom_time_series.R
\name{plot_hom_series}
\alias{plot_hom_series}
\title{Visualize homogenized time series from Climatol}
\usage{
plot_hom_series(
  id,
  brks = break_points,
  h_series = hom_series,
  o_series = series_comb
)
}
\arguments{
\item{id}{staid of the station}

\item{brks}{break points from climatol}

\item{h_series}{homogenized time series}

\item{o_series}{original series from ECAD}
}
\description{
plots the homogenized time series with the interactive plotly library.
}
