% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ecad_climatol.R
\name{prepare_ecad_climatol}
\alias{prepare_ecad_climatol}
\title{Prepare ECAD data for homogenization with climatol}
\usage{
prepare_ecad_climatol(t.start, t.stop, stations_qq, series_qq)
}
\arguments{
\item{t.start}{starting date to be homogenized}

\item{t.stop}{end date to be homogenized}

\item{stations_qq}{station data}

\item{series_qq}{global radiation observations}
}
\description{
Organizes the station data and time-series for homogenization
with the climatol package. The .dat and .est files are saved in the current work directory.
}
\author{
Marieke Dirksen
}
