% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ecad.R
\name{prepare_series_ss2qq}
\alias{prepare_series_ss2qq}
\title{Prepare monthly files of global radiation from sunshine duration}
\usage{
prepare_series_ss2qq(main_dir)
}
\arguments{
\item{main_dir}{main path where the SunCloud folder with all the data is}
}
\description{
Reads the monthly ss2qq files
}
\author{
Marieke Dirksen
}
