% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ecad.R
\name{read_ECA_info}
\alias{read_ECA_info}
\title{Read metadata info from ECAD}
\usage{
read_ECA_info(fname)
}
\arguments{
\item{fname}{file name of the metadata}
}
\value{
returns a dataframe with lat,lon converted to decimal degrees
}
\description{
reads the metadata files from ECAD with a selection of colomns
}
\author{
Marieke Dirksen
}
