% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ecad.R
\name{read_monthly_QQ}
\alias{read_monthly_QQ}
\title{Read monthly qq}
\usage{
read_monthly_QQ(
  monthly_qq = paste0(file_loc$main_loc, "/", file_loc$global_radiation,
    "/QQ_monthly_cs.txt")
)
}
\arguments{
\item{monthly_qq}{name of the file}
}
\description{
Reads the monthly qq files fields month_year, QQm and STAID calculated by \link{clearsky_monthly_qq_cc},
\link{clearsky_monthly_qq_cc} and \link{fill_time_series}.
}
\author{
Marieke Dirksen
}
