% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ecad.R
\name{read_qq}
\alias{read_qq}
\title{Read qq}
\usage{
read_qq(file_qq)
}
\arguments{
\item{file_qq}{global radiation file}
}
\value{
returns a dataframe with dates converted into date format
}
\description{
Read the qq file which is used in \link{allsky_monthly_qq}.
}
\author{
Marieke Dirksen
}
