% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ecad.R
\name{read_qq_cc}
\alias{read_qq_cc}
\title{Read and combine qq and cc files}
\usage{
read_qq_cc(file_qq, file_cc)
}
\arguments{
\item{file_qq}{global radiation file}

\item{file_cc}{cloud cover file, matching with file_qq}
}
\value{
a combined dataframe with dates converted to date format
}
\description{
reads qq and cc and merges both files with matching STAID and DATE, used for
\link{clearsky_monthly_qq_cc}.
}
\author{
Marieke Dirksen
}
