% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_averages.R
\name{seasonal_trend}
\alias{seasonal_trend}
\title{Seasonal trends}
\usage{
seasonal_trend(qq_month)
}
\arguments{
\item{qq_month}{monthly global radiation observations}
}
\value{
returns a dataframe with seasonal and normalized seasonal values
}
\description{
Calculates seasonal trends from monthly averaged values calculated from
\link{clearsky_monthly_qq_cc} and \link{allsky_monthly_qq}.
}
\author{
Marieke Dirksen
}
