% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_analysis.R
\name{fit_sens_slope}
\alias{fit_sens_slope}
\title{Fit sens slope to the series}
\usage{
fit_sens_slope(series)
}
\arguments{
\item{series}{time-series to be analysed}
}
\description{
Fits sens slope to a time series within a period. It is adviced to run
\link{get_change_point} first to find a proper time period for the fitting (without change points).
}
\note{
See also \url{https://rcompanion.org/handbook/G_10.html}
}
\author{
Marieke Dirksen
}
